requirements.txt
discord.py>=2.3.0
discord.py-self>=2.1.0
flask>=3.0.0
gunicorn>=21.0.0
requests
PyNaCl>=1.5.0
yt-dlp
colorama
